requirements-dev.txt
apipkg==1.5
appnope==0.1.0
argh==0.26.2
asn1crypto==0.24.0
async-lru==0.1.0
asyncio-cancel-token==0.1.0a2
atomicwrites==1.2.0
attrdict==2.0.0
attrs==18.1.0
backcall==0.1.0
bloom-filter==1.3
bumpversion==0.5.3
cachetools==2.1.0
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
coincurve==8.0.2
colorama==0.3.9
coverage==4.5.1
cryptography==2.3.1
cytoolz==0.9.0.1
decorator==4.3.0
Django==2.1.1
django-filter==2.0.0
django-allauth==0.37.1
django-filter==2.0.0
django-rest-auth==0.9.3
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
docopt==0.6.2
eth-abi==1.2.0
eth-account==0.3.0
eth-bloom
eth-hash
eth-keyfile
eth-keys==0.2.0b3
eth-rlp==0.1.2
eth-tester==0.1.0b31
eth-typing==1.3.0
eth-utils==1.2.0
execnet==1.5.0
flaky==3.4.0
hexbytes==0.1.0
httplib2==0.9.2
hypothesis==3.69.9
idna==2.7
ipython==6.5.0
ipython-genutils==0.2.0
jedi==0.12.1
keyring==10.4.0
keyrings.alt==2.2
launchpadlib==1.10.5
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lru-dict==1.1.6
lxml==4.2.4
Mako==1.0.7
Markdown==2.6.11
MarkupSafe==1.0
more-itertools==4.3.0
netdisco==2.0.0
netifaces==0.10.7
oauth==1.0.1
olefile==0.44
parsimonious==0.8.0
parso==0.3.1
path.py==7.7.1
pathtools==0.1.2
pbr==4.2.0
pexpect==4.6.0
pickleshare==0.7.4
Pillow==4.1.1
pluggy==0.7.1
plyvel==1.0.5
prompt-toolkit==1.0.15
ptyprocess==0.6.0
py==1.6.0
py-ecc==1.4.3
py-evm==0.2.0a32
py-geth==2.0.1
py-solc==3.1.0
pycairo==1.17.1
pycparser==2.18
git+https://github.com/Legrandin/pycryptodome
pyethash==0.1.27
Pygments==2.2.0
pysha3==1.0.2
pytest==3.7.3
pytest-forked==0.2
pytest-mock==1.10.0
pytest-pythonpath==0.7.3
pytest-watch==4.2.0
pytest-xdist==1.23.0
python-dateutil==2.7.3
python-debian==0.1.30
pytz==2018.5
pyxdg==0.25
PyYAML==3.13
reportlab==3.4.0
requests==2.19.1
rlp==1.0.2
SecretStorage==2.3.1
semantic-version==2.6.0
simplegeneric==0.8.1
simplejson==3.11.1
six==1.11.0
testresources==2.0.1
toolz==0.9.0
tox==3.2.1
tqdm==4.25.0
traitlets==4.3.2
trie==1.3.8
uPnPClient==0.0.8
urllib3==1.23
virtualenv==16.0.0
wadllib==1.3.2
watchdog==0.8.3
wcwidth==0.1.7
web3
websockets==6.0
when-changed==0.3.0
zeroconf==0.20.0
zope.interface==4.3.2
psycopg2==2.7.5
scrypt==0.8.6